requirements.txt
Flask
requests
validators
selenium
webdriver-manager
pyopenssl
ping3
requests
sql
flask_sqlalchemy
mysql-connector-python
